backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
pydantic>=2.7.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2
